+++ src/oihana/arango/clients/graph/Graph.php
@@ -7,8 +7,8 @@ use oihana\enums\http\HttpMethod ;
 use oihana\arango\clients\Database ;
 use oihana\arango\clients\enums\ArangoRoute ;
 use oihana\arango\clients\exceptions\ArangoException ;
-use oihana\arango\clients\exceptions\HttpException ;
 
+use function oihana\arango\clients\helpers\probeExists ;
 use function oihana\arango\clients\helpers\unwrapField ;
 
 /**
@@ -325,19 +325,7 @@ readonly class Graph
      */
     public function exists() : bool
     {
-        try
-        {
-            $this->database->request( method : HttpMethod::GET , path : $this->path() ) ;
-            return true ;
-        }
-        catch ( HttpException $e )
-        {
-            if ( $e->getCode() === 404 )
-            {
-                return false ;
-            }
-            throw $e ;
-        }
+        return probeExists( fn() => $this->database->request( HttpMethod::GET , $this->path() ) ) ;
     }
 
     /**