+++ src/oihana/arango/clients/view/View.php
@@ -7,9 +7,10 @@ use oihana\enums\http\HttpMethod ;
 use oihana\arango\clients\Database ;
 use oihana\arango\clients\enums\ArangoRoute ;
 use oihana\arango\clients\exceptions\ArangoException ;
-use oihana\arango\clients\exceptions\HttpException ;
 use oihana\arango\clients\view\enums\ViewField ;
 use oihana\arango\clients\view\enums\ViewType ;
+
+use function oihana\arango\clients\helpers\probeExists ;
 
 /**
  * Operations scoped to a single ArangoSearch view on the server.
@@ -219,19 +220,7 @@ readonly class View
      */
     public function exists() : bool
     {
-        try
-        {
-            $this->database->request( method : HttpMethod::GET , path : $this->path() ) ;
-            return true ;
-        }
-        catch ( HttpException $e )
-        {
-            if ( $e->getCode() === 404 )
-            {
-                return false ;
-            }
-            throw $e ;
-        }
+        return probeExists( fn() => $this->database->request( HttpMethod::GET , $this->path() ) ) ;
     }
 
     /**